Independent SwipeEventArgs File
