+++ src/MudBlazor.UnitTests/Components/SwipeTests.cs
@@ -33,11 +33,27 @@ namespace MudBlazor.UnitTests.Components
             var comp = Context.RenderComponent<SwipeAreaOnSwipeEndTest>();
             var swipe = comp.FindComponent<MudSwipeArea>();
 
+            // Swipe below the sensitivity should not make change.
             var initialTouchPoints = new TouchPoint[]
             {
                 new TouchPoint() {ClientX = 0, ClientY = 0},
             };
             var touchPoints = new TouchPoint[]
+            {
+                new TouchPoint() {ClientX = 20, ClientY = 20},
+            };
+
+            await comp.InvokeAsync(() => swipe.Instance.OnTouchStart(new TouchEventArgs() { Touches = initialTouchPoints }));
+            await comp.InvokeAsync(() => swipe.Instance.OnTouchEnd(new TouchEventArgs() { ChangedTouches = touchPoints }));
+
+            comp.WaitForAssertion(() => comp.Instance.SwipeDirection.Should().Be(SwipeDirection.None));
+            comp.WaitForAssertion(() => comp.Instance.SwipeDelta.Should().Be(null));
+
+            initialTouchPoints = new TouchPoint[]
+            {
+                new TouchPoint() {ClientX = 0, ClientY = 0},
+            };
+            touchPoints = new TouchPoint[]
             {
                 new TouchPoint() {ClientX = 150, ClientY = 200},
                 new TouchPoint() {ClientX = 100, ClientY = 50},
@@ -46,8 +62,8 @@ namespace MudBlazor.UnitTests.Components
             await comp.InvokeAsync(() => swipe.Instance.OnTouchStart(new TouchEventArgs() { Touches = initialTouchPoints }));
             await comp.InvokeAsync(() => swipe.Instance.OnTouchEnd(new TouchEventArgs() { ChangedTouches = touchPoints }));
 
-            comp.WaitForAssertion(() => comp.Instance._swipeDirection.Should().Be(SwipeDirection.TopToBottom));
-            comp.WaitForAssertion(() => comp.Instance._swipeDelta.Should().Be(-200));
+            comp.WaitForAssertion(() => comp.Instance.SwipeDirection.Should().Be(SwipeDirection.TopToBottom));
+            comp.WaitForAssertion(() => comp.Instance.SwipeDelta.Should().Be(-200));
         }
 
         [Test]
